requirements.txt
fastapi
uvicorn[standard]
sse-starlette
anthropic
openai
akshare
pandas
numpy
plotly
python-dotenv
jinja2
python-multipart
baostock
yfinance
